LPD-57376 - Adjust types
