+++ modules/apps/frontend-data-set/frontend-data-set-web/src/main/resources/META-INF/resources/utils/actionItems/triggerAction.ts
@@ -5,8 +5,8 @@
 
 import {navigate} from 'frontend-js-web';
 
+import {ICreationActionItem} from '../../';
 import {IFrontendDataSetContext} from '../../FrontendDataSetContext';
-import {ICreationActionItem} from '../../management_bar/controls/CreationMenu';
 import {OPEN_MODAL, OPEN_SIDE_PANEL} from '../../utils/eventsDefinitions';
 import {resolveModalSize} from '../../utils/modals/resolveModalSize';
 import {ACTION_ITEM_TARGETS} from './constants';
