+++ modules/apps/object/object-web/src/main/resources/META-INF/resources/js/components/ViewObjectDefinitions/ViewObjectDefinitions.tsx
@@ -31,7 +31,7 @@ import {
 import './ViewObjectDefinitions.scss';
 import {ViewObjectDefinitionsModals} from './ViewObjectDefinitionsModals';
 
-import type {IItemsActions} from '@liferay/frontend-data-set-web';
+import type {ICreationActionItem, IItemsActions} from '@liferay/frontend-data-set-web';
 
 import type {FDSItem, IFDSTableProps} from '../../utils/fds';
 
@@ -52,7 +52,7 @@ interface ViewObjectDefinitionsProps extends IFDSTableProps {
 	modelBuilderURL: string;
 	nameMaxLength: string;
 	objectDefinitionsCreationMenu: {
-		primaryItems?: any[];
+		primaryItems: Array<ICreationActionItem>;
 		secondaryItems?: any[];
 	};
 	objectDefinitionsFDSActionDropdownItems: IItemsActions[];
